fix(utils): domain suffix improperly checked
This fixes an issue where a domain of example.com would be considered to be a suffix of the URL https://app.notexample.com.
