+++ internal/utils/url.go
@@ -1,8 +1,10 @@
 package utils
 
 import (
+	"fmt"
 	"net/url"
 	"path"
+	"strings"
 )
 
 // URLPathFullClean returns a URL path with the query parameters appended (full path) with the path portion parsed
@@ -26,4 +28,32 @@ func URLPathFullClean(u *url.URL) (output string) {
 	default:
 		return path.Clean(u.Path)
 	}
+}
+
+// URLDomainHasSuffix determines whether the uri has a suffix of the domain value.
+func URLDomainHasSuffix(uri url.URL, domain string) bool {
+	if uri.Scheme != https {
+		return false
+	}
+
+	if uri.Hostname() == domain {
+		return true
+	}
+
+	if strings.HasSuffix(uri.Hostname(), period+domain) {
+		return true
+	}
+
+	return false
+}
+
+// IsRedirectionURISafe determines whether the URI is safe to be redirected to.
+func IsRedirectionURISafe(uri, protectedDomain string) (safe bool, err error) {
+	var parsedURI *url.URL
+
+	if parsedURI, err = url.ParseRequestURI(uri); err != nil {
+		return false, fmt.Errorf("failed to parse URI '%s': %w", uri, err)
+	}
+
+	return parsedURI != nil && URLDomainHasSuffix(*parsedURI, protectedDomain), nil
 }
